Don't use <stdint.h> when targeting strict ISO C on GNU toolchains
Up to now we unconditionally included the C99 <stdint.h> header
and assumed the 'long long' type would be available everywhere,
making the '%ll' format specifier the default for 64-bit ints.

All safe and correct assumptions in this day and age, except it
prevents consuming programs from being strict ANSI C.

For those using GNU compilers (GCC and, especially, MinGW), this
guards the inclusion of <stdint.h> and provides polyfill for the
int types we need.

There are *no* plans to adapt this hack for MSVC, which didn't
even define __STDC_VERSION__ before version 16.6 of Visual Studio
2019 [1]. Apple Clang pulls in the standard int types from a
collection of BSD headers and should not be affected in any case.

----
[1] https://docs.microsoft.com/en-us/cpp/overview/visual-cpp-language-conformance?view=msvc-160#note_C